Fix coverity warning
CID 332936 (#1 of 1): Result is not floating-point (UNINTENDED_INTEGER_DIVISION)
integer_division: Dividing integer expressions size and 2, and then converting
the integer quotient to type float. Any remainder, or fractional part of the
quotient, is ignored.

+++ components/detournavigator/recastmeshbuilder.cpp
@@ -36,7 +36,7 @@ namespace DetourNavigator
 
         TileBounds maxCellTileBounds(int size, const osg::Vec3f& shift)
         {
-            const float halfCellSize = size / 2;
+            const float halfCellSize = static_cast<float>(size) / 2;
             return TileBounds {
                 osg::Vec2f(shift.x() - halfCellSize, shift.y() - halfCellSize),
                 osg::Vec2f(shift.x() + halfCellSize, shift.y() + halfCellSize)
